requirements.txt
pypresence==4.6.1
python-dotenv==1.2.2
